refactor: 하드 코딩값 상수 선언
