+++ src/car.js
@@ -2,12 +2,13 @@ export default class Car {
   constructor(name) {
     this.name = name;
 
+    this.MOVE_CONDITION = 4;
     this.distance = 0;
     this.position = "";
   }
 
   move(number) {
-    if (number >= 4) {
+    if (number >= this.MOVE_CONDITION) {
       this.distance++;
       this.position += "-";
     }
