+++ src/message.js
@@ -1,0 +1,5 @@
+export const alertMessage = {
+  INVALID_LENGTH_MESSAGE: "자동차 이름은 한 글자 이상 다섯 글자 이하여야 합니다.",
+  DUPLICATE_NAME_MESSAGE: "자동차 이름은 중복될 수 없습니다.",
+  INVALID_NUMBER_MESSAGE: "1이상의 자연수만 입력할 수 있습니다.",
+};
